fix: set CMAKE_CXX_STANDARD to 20 and fix atomic narrowing for macOS
LLVM 22 auto-detects c++26 which has header incompatibilities with
Xcode 16.4 SDK. Pin to c++20 which the project actually uses.

Also cast atomic_t initializers to uint64_t to fix narrowing conversion
errors in C++ compilation (-Wc++11-narrowing).

+++ include/ascii-chat/log/log.h
@@ -29,12 +29,12 @@
 // atomic_t is available in both C and C++ contexts
 #define LOG_ATOMIC_UINT64 atomic_t
 #ifdef NDEBUG
-#define LOG_ATOMIC_UINT64_INIT(val) {.impl = 0}
+#define LOG_ATOMIC_UINT64_INIT(val) {.impl = (uint64_t)0}
 #else
 // In debug mode, atomic_t has debug fields that need initialization
-#define LOG_ATOMIC_UINT64_INIT(val) {.impl = 0, .last_store_time_ns = 0, .last_load_time_ns = 0, \
-                                      .store_count = 0, .load_count = 0, .cas_count = 0, \
-                                      .cas_success_count = 0, .fetch_count = 0}
+#define LOG_ATOMIC_UINT64_INIT(val) {.impl = (uint64_t)0, .last_store_time_ns = (uint64_t)0, .last_load_time_ns = (uint64_t)0, \
+                                      .store_count = (uint64_t)0, .load_count = (uint64_t)0, .cas_count = (uint64_t)0, \
+                                      .cas_success_count = (uint64_t)0, .fetch_count = (uint64_t)0}
 #endif
 #include <stdbool.h>
 #include <stdint.h>
